+++ package.json
@@ -14,7 +14,6 @@
         "geohash-distance": "*",
         "gun": "^0.9.999996",
         "gun-tag": "*",
-        "helmet": "^3.15.0",
         "latlon-geohash": "*",
         "leaflet": "^1.4.0",
         "leaflet.locatecontrol": "^0.66.0",
